- Bug fixes - String handing removed from non-x86 parts of code - Changes to register, deregister and register update function API's
git-svn-id: svn+ssh://192.168.0.9/home/common-files/subversion/devel/CoAP/nsdl-c-simple@15014 b4939478-c410-0410-9151-a6aaa495c60e

+++ libCoap/sn_coap_protocol_ietf_draft_09.h
@@ -21,7 +21,7 @@
 #define RESPONSE_TIMEOUT        2   /* Initial resending timeout as seconds, value is specified in IETF CoAP specification */
 #define RESPONSE_RANDOM_FACTOR  1   /* Resending random factor, value is specified in IETF CoAP specification */
 
-#define COAP_SEND_MSG_ONCE      0xFFFFFFFF /* Sends CoAP message once from Linked list, no resendins */
+#define COAP_SEND_MSG_ONCE      ((uint32_t)0xFFFFFFFF) /* Sends CoAP message once from Linked list, no resendins */
 
 /* * * * * * * * * * * * * * */
 /* * * * ENUMERATIONS  * * * */
